+++ rbac/management/role/view.py
@@ -24,10 +24,10 @@ from django.db.models.aggregates import Count
 from django.http import Http404
 from django.utils.translation import gettext as _
 from django_filters import rest_framework as filters
+from management.filters import CommonFilters
 from management.permissions import RoleAccessPermission
 from management.querysets import get_role_queryset
 from management.role.serializer import AccessSerializer, RoleDynamicSerializer
-from management.utils import validate_and_get_key
 from rest_framework import mixins, serializers, viewsets
 from rest_framework.decorators import action
 from rest_framework.filters import OrderingFilter
@@ -38,8 +38,6 @@ from .serializer import RoleSerializer
 TESTING_APP = os.getenv("TESTING_APPLICATION")
 APP_WHITELIST = ["cost-management", "remediations"]
 ADDITIONAL_FIELDS_KEY = "add_fields"
-NAME_MATCH_KEY = "name_match"
-VALID_NAME_MATCHES = ["partial", "exact"]
 VALID_FIELD_VALUES = ["groups_in_count", "groups_in"]
 LIST_ROLE_FIELDS = [
     "uuid",
@@ -58,17 +56,8 @@ if TESTING_APP:
     APP_WHITELIST.append(TESTING_APP)
 
 
-class RoleFilter(filters.FilterSet):
+class RoleFilter(CommonFilters):
     """Filter for role."""
-
-    def name_filter(self, queryset, field, value):
-        """Filter for role to lookup name, partial or exact."""
-        match_criteria = validate_and_get_key(self.request.query_params, NAME_MATCH_KEY, VALID_NAME_MATCHES, "partial")
-
-        if match_criteria == "partial":
-            return queryset.filter(name__icontains=value)
-        elif match_criteria == "exact":
-            return queryset.filter(name__iexact=value)
 
     name = filters.CharFilter(field_name="name", method="name_filter")
